feat(plugin): adapter plugin -- db2
