+++ configure/etc/conf/plugins.properties
@@ -9,6 +9,7 @@ datacap-plugin-kafka=plugin/datacap-plugin-kafka/pom.xml
 datacap-plugin-redis=plugin/datacap-plugin-redis/pom.xml
 datacap-plugin-h2=plugin/datacap-plugin-h2/pom.xml
 datacap-plugin-hdfs=plugin/datacap-plugin-hdfs/pom.xml
+datacap-plugin-db2=plugin/datacap-plugin-db2/pom.xml
 datacap-convert-txt=convert/datacap-convert-txt/pom.xml
 datacap-convert-json=convert/datacap-convert-json/pom.xml
 datacap-convert-xml=convert/datacap-convert-xml/pom.xml
