--- plugin/datacap-jdbc-db2/src/main/java/io/edurt/datacap/plugin/jdbc/db2/Db2Adapter.java
@@ -1,13 +1,0 @@
-package io.edurt.datacap.plugin.jdbc.db2;
-
-import io.edurt.datacap.spi.adapter.JdbcAdapter;
-import io.edurt.datacap.spi.connection.JdbcConnection;
-
-public class Db2Adapter
-        extends JdbcAdapter
-{
-    public Db2Adapter(JdbcConnection jdbcConnection)
-    {
-        super(jdbcConnection);
-    }
-}
