+++ plugin/datacap-plugin-db2/src/main/java/io/edurt/datacap/plugin/Db2Plugin.java
@@ -1,5 +1,5 @@
 package io.edurt.datacap.plugin;
 
-public class H2Plugin
+public class Db2Plugin
         extends Plugin
 {}
